Misc updates
* Change version in Makefile to point `dev` to prevent accidentaly
pushing to an already tagged version
* Added a new `-dev` version for testing. By default the master is
always in `dev` version.
* Added a new `-latest-stable.yaml` file that always points to the
latest stable release. This will be updated whenever we release a new
version.
* Removed unused files
* Update README to point to the latest stable release

I'm going to update all these in upcoming PR's to fetch a version from a
hardcoded place and use `ldflags` to generate and update all `dev`
occurences with a new version. Basically there will be some scripts and
tooling to make updating easier going forward. This PR just laids out
the foundation and improves the current situation.