New release v0.3.0
This release is intended to be used with Kubernetes `v1.12.x` and is not
compatible with older versions of Kubernetes. The latest CSI changes in
`v1.12.x` are not compatible with older version unfortunately, therefore
going forward we will not support older version anymore. Anyone who
wants to use this plugin is free to use the
[v0.2.0](https://github.com/digitalocean/csi-digitalocean/blob/master/deploy/kubernetes/releases/csi-digitalocean-v0.2.0.yaml)
version.