csi: fix incosistent driver name usages
Our driver name has changed from `com.digitalocean.csi.dobs` to
`dobs.csi.digitalocean.com` because of the upcoming changes to the CSI
spec that decided to follow a forward domain notation instead of reverse
domain notation.

Seems like the names were incorrect in many places and we also forget to
update it in various places. I made sure to export `driver.DriverName`
and use that variable everywhere in the code. So in the future, if we
need to change it again, we'll make sure it's updated correctly from a
single place.