Return InvalidArgument instead of AlreadyExists for invalid VolumeCapabilities
From reading through the Error Scheme section[1] of the spec and
comparing that to the list of CreateVolume Errors[2], I came to the
conclusion that AlreadyExists does not quite apply as outlined in the
RPC-specific error code section, and instead we should be referring to
the general Error Scheme.

This changes the error codes around to be more appropriate for signaling
what the actual error was, and thus better aligning with the error
description we pass back.

[1]: https://github.com/container-storage-interface/spec/blob/master/spec.md#error-scheme
[2]: https://github.com/container-storage-interface/spec/blob/master/spec.md#createvolume-errors