csi: v1.0.0
CSI v1.0.0 is released (\o/): https://github.com/container-storage-interface/spec/releases/tag/v1.0.0.

With that, there were many changes and updates so this PR combines all
these changes.

* Create a default `VolumeSnapshotClass` so people don't have to define
them when they use DO for their snapshotting needs.
* Add new `csi-snapshotter` sidecar that is responsible of calling the
three CSI methods we implemented: `ListSnapshots`, `CreateSnapshot` and
`DeleteSnapshot`. In addition to this, the `csi-snapshotter` sidecar is
also responsible of creating the `VolumeSnapshotClasss` CRD
* Updated sidecars to the newest and latest
* Added back the ClusterRole's for each sidecar. Kubernetes no more
carries them and it's up to the sidecars to provider them to the users.