deploy: add VolumeSnapshot CRD's
The following CRD's are created by the csi-snapshotter, however it
complicates installing a driver, because we're not able to install a custom
VolumeSnapshotClass until the csi-snapshotter sidecar is up and running.

We pulled out the CRD's and put them here to simplify the installation
for the users. We need to make sure these are up to date with the
original ones whenever we release a new version:
https://github.com/kubernetes-csi/external-snapshotter/blob/master/cmd/csi-snapshotter/create_crd.go