driver: don't check limits on worker nodes
We're checking the volume limit during the driver startup. However
because the driver is used both for the controller and node part, it
fails during the node startup because we're deploying the driver to the
nodes without a token. The DO token is only needed by the controller.

The `checkLimit()` function would therefore print when deployed to the
nodes:

```
time="2019-01-14T20:13:52Z" level=warning msg="CSI plugin will not
function correctly, please resolve volume limit" error="rpc error: code
= Internal desc = couldn't get account information to check volume
limit: GET https://api.digitalocean.com/v2/account: 401 Unable to
authenticate you."
```

This PR fixes that to not to call `checkLimit` if the token is not
passed, which is a valid case for drivers deployed to the worker nodes.

supersedes #127