Allow a custom driver name to be specified on the commandline
The name of our CSI driver changed between the 0.2 and 0.4 releases,
from `com.digitalocean.csi.dobs` to `dobs.csi.digitalocean.com`. This
means that when upgrading a cluster from Kubernetes 1.11 (which uses CSI
plugin 0.2) to Kubernetes 1.12 (which uses CSI plugin 0.4) volumes
created before the upgrade have the old driver name in various
annotations and spec fields. This prevents the new driver from managing
the volume, leading to broken workloads.

To handle the upgrade case, allow a custom driver name to be specified
on the commandline. This way, a cluster that was once running 1.11 and
thus uses the old name can continue using the old name. By default the
new name will be used.