test: Add kustomization for deploying a dev version
In order to test dev versions of the CSI driver on DOKS clusters, we
need to deploy a second storage class that uses an updated version of
the CSI driver alongside the built-in version.

Add a [kustomize](https://github.com/kubernetes-sigs/kustomize) config
that can transform our dev manifest to be used for this purpose. This
involves renaming everything (appending `-dev` to the storage class,
deployment, statefulset, etc. names) and running the CSI driver with the
`--driver-name` argument.