test: Add a script and instructions for testing on DOKS
Add a script that uses `kustomize` to deploy a dev version of the CSI
driver to a DOKS cluster alongside the pre-installed version. This makes
it easy to use a DOKS cluster for testing changes to the CSI driver.

Update the README and add a README for the test version deployment
process.