Add v0.2.1 manifest to master
The v0.2.1 manifest was only available in the release-0.2.0 branch;
however, we always reference manifests in master even from release
branches.

We may want to change the approach and reference local to a branch or
tag going forward. For now though, we fix the issue in a manner that is
consistent with our current approach.