Refactor waitAction
This change refactors waitAction to use wait.PollUntil from the upstream
apimachinery library to hide the underlying polling infrastructure and
simplify the implementation.

We also exit the method if the context got cancelled while waiting for a
response from the volume API. Previously, we would iterate through the
for loop once again and possibly do another unnecessary API interaction
if the ticker was ready again by that time.

Finally, we add a unit test (which required making the wait action
timeout injectable).