Prevent kustomize modifications from polluting git
We support injecting a different plugin image into the dev/latest
manifests. However, this pollutes git my updating the VCS-controlled
kustomization.yaml file.
This change sets up a bash trap to undo any image changes.

As a drive-by improvement, we also simplify the name of the image to
do-csi-plugin since it is supposed to reference a generic name and not a
specific image version (which is set by 'kustomize edit set image').