Check volume exists during NodeUnpublishVolume
The CSI spec mandates that we return NOT_FOUND if the volume isn't found
[1]. Otherwise, the calling sidecar will go into an endless retry loop
and block other operations on the volume.

A test for this case should go into github.com/kubernetes-csi/csi-test
but was yet missing. A PR has been submitted [2] and this change has
been tested against it. As soon as upstream approves, we can follow up
with another PR to incorporate an updated csi-sanity package.

[1]: https://github.com/container-storage-interface/spec/blob/4731db0e0bc53238b93850f43ab05d9355df0fd9/spec.md#nodeunpublishvolume-errors
[2]: https://github.com/kubernetes-csi/csi-test/pull/242