Support checkLimit for multiple pages (#235)
This change uses the Meta.Total field from the godo response when
listing volumes for account limit checking, thereby supporting limits
beyond the size of a single page.

We also refactor checkLimits to not return gRPC codes but leave that to
the calling function. This also allows us to distinguish between true
errors and limit violations, and better handle each case.