Revert "Check volume exists during NodeUnpublishVolume"
After consulting with SIG Storage [1], it turned out that the kind of
volume existence check implemented by the reverted commit is to be a
node-local operation and not something we should reach out to the API
for.
Additonally, it primarily applies to volume types that need to connect
to the device locally, such as local disks. This is not the case for our
type of storage.

This reverts commit d009d46792da41948a6a3ebf27b8ce52988fce00.

[1]: https://kubernetes.slack.com/archives/C8EJ01Z46/p1576538436024100