Remove/generalize references to latest plugin version
Pointing at and keeping track of the latest plugin version in the README
file has been the source of two kinds of problems:

1. The README file is being updated during development, with changes
   possibly relating to upcoming features. Mentioning the latest release
   in the same file has led to the wrong assumption with some users that
   references to yet unreleased features are already available with the
   latest release. This is because we conflate changes pertaining to
   master only with references to tags / stable identifiers.
2. Bumping/replacing the version in a purely text-based manner is
   error-prone and has already caused issues in our manifest files where
   we apply similar patterns [1].

Pointing at the latest release in the README seems unnecessary since
that is what the Github Releases page is for; additionally,
version-specific documentation should be retrieved via tag-based links.
Consequently, this change removes / generalizes the references to
improve clarity and simplify house-keeping chores.

[1]: https://github.com/digitalocean/csi-digitalocean/pull/198