Document all supported Kubernetes versions (#251)
This change explicitly mentions all Kubernetes releases we officially
support with our CSI driver. It seems preferable to the previous style
which leaves uncertainty as to whether the most recently published
Kubernetes release is already supported.

We also drop the `dev` row as it is fairly uninteresting to
non-developers and is easy to forget to be updated.