Fix ListSnapshots paging
This changes provides the following fixes and improvements to
ListSnapshots:

- Use paging to collect snapshots beyond the first page. Previously, we
  would only return snapshots from the first page.
- Handle StartingToken and MaxEntries such that we use paging
  efficiently and skip initial, unneeded snapshots.
- Extend fake snapshot driver to support paging.
- Add tests.

Note that Kubernetes / the csi-snapshotter sidecar currently do not
invoke ListSnapshots without the snapshot ID parameter, which means that
the fixed code is not executed in production. However, it is used by
csi-test / the sanity package, and other COs (Container Orchestrators)
may potentially use it as well as Kubernetes going forward.