Enable tests for node volume attachment limits
The csi-test sanity package ships with off-by-default tests to validate
per-node attachment limits. This change toggles the corresponding test
configuration flag to enable the tests.

The change requires modifying our fake driver to return a 422 HTTP error
code when the limit is exceeded. As a consequence, we also need to
customize the IdempotentCount test setting which parameterizes the
'should be idempotent' test that creates the given number of volumes in
sequence. The default value of 10 causes our (fake) limit to be
exceeded, which is why we tune it down to 5.

The test also revealed that we missed to handle the case where the node
volume attachment limit is exceed during ControllerPublishVolume. We
extend our error handling to identify this case and return an
RESOURCE_EXHAUSTED code accordingly.