e2e: handle 429 TooManyRequests error
During the polling for the cluster create request to finish, we
frequently fail because we have used up all of our API quota budget.
In particular, this can happen when multiple e2e tests are running
concurrently.

This change addresses the problem why handling a seen 429 with some
extra sleep time to replenish our API request budget.

We also bump the involved timeouts to account for the extra test time we
might need now.