Rename csi-digitalocean-latest to csi-digitalocean-dev
A number of users have been assuming that the -latest suffix in our
Kubernetes releases directory implies a production-ready collection of
manifests, whereas it is meant to be the manifests from the latest
commit (and, as such, possibly not quite stable).

To make things more clear, we rename the suffix to -dev.