Fix built-in release version
In a previous change, we started to release container images by
promoting/retagging the latest :master image. However, this did not
yield an updated version string baked into the binary; instead, we ended
up with the "latest" version used when we build the latest master.

This change modifies the release Github Action by recompiling and
packaging the binary so that the right version is included.