Update csi-provisioner to v2.0.2 and csi-snapshotter to v3.0.0
csi-provisioner v2.0.2 brings a small watch optimization that is
actually not relevant for. However, it is still good to be on the latest
version.

csi-snapshotter v3.0.0 brings stricter validation and several bugfixes.
The breaking changes pertain to Go interface adjustments (e.g.,
snapshotter using context when talking to the API server) which do not
affect us since we consume the binary only.