Allow releasing from HEAD in Github Actions
Our Makefile contains an (ugly) chain of conditions to prevent releasing
anything but the "dev" tag from a non-master branch. When doing releases
from Github Actions, however, the checked out repo is at HEAD and thus
qualifies to hit the case we protect against.

We solve the matter by introducing an overriding environment variable
and setting in the release Workflow.