Allow releasing from tag in Github Actions (#358)
Our Makefile contains an (ugly) chain of conditions to prevent releasing
anything but the dev tag from a non-master branch. This breaks our release
after we moved to rebuilding instead of just retagging (#342).

We solve the matter by requiring a semver $VERSION in the push target,
which is exactly what our release action already provides.