Set default fstype to fix fsGroup usage
The provisioner sidecar v2 dropped setting a default file-system type.
When we upgraded, we did not set it explicitly because our driver
already handles the case where no file-system is specified. As it turns
out, however, the kubelet gates the application of the fsGroup security
context parameter setting the file permissions of a volume on the
availability of a file-system at the time the PV is created and
otherwise ignores it.

We restore proper functionality of fsGroup by setting the default
file-system type to ext4.