allow running the CSI controller outside DigitalOcean infrastructure
When running the CSI controller outside DO, it will still attempt to contact the metadata service due to it attempting to collect the droplet ID.

When running in controller mode, it will ignore the error from DropletID(). With this change, it fixes the "Controller only mode not in DigitalOcean" mentioned in the README file.

Signed-off-by: Thomas Kooi <t.j.kooi@avisi.nl>