Disambiguate non-existent volumes from ongoing detachments
When a detachment is issued, a 404 response may be returned that does
not actually indicate a non-existent volumes but implies that the
detachment is not yet known to the Storage backend. Disambiguate by
checking the droplet's attached volumes and assume a detachment to be
pending if the volume is still found.